Use size_t indexes for null-terminated arrays
When operating on arrays that don't have a specified integer bound,
use size_t indexes for improved safety.  Reported by James Watt.

+++ src/kdc/main.c
@@ -611,7 +611,7 @@ initialize_realms(krb5_context kcontext, int argc, char **argv,
     const char          *hierarchy[3];
     char                *no_referral = NULL;
     char                *hostbased = NULL;
-    int                  db_args_size = 0;
+    size_t               db_args_size = 0;
     char                **db_args = NULL;
 
     extern char *optarg;
